+++ ports/arm-cm/qxk/iar/qxk_port.h
@@ -40,19 +40,19 @@
 #define QXK_PORT_H
 
 /* determination if the code executes in the ISR context */
-#define QXK_ISR_CONTEXT_() (__get_IPSR() != (uint32_t)0)
+#define QXK_ISR_CONTEXT_() (__get_IPSR() != 0U)
 
 #if (__CORE__ == __ARM6M__)  /* Cortex-M0/M0+/M1? */
 
     #define QXK_CONTEXT_SWITCH_() \
-        (*Q_UINT2PTR_CAST(uint32_t, 0xE000ED04U) = (uint32_t)(1U << 28))
+        (*Q_UINT2PTR_CAST(uint32_t, 0xE000ED04U) = (1U << 28))
 
 #else /* Cortex-M3/M4/M7 */
 
-    #define QXK_CONTEXT_SWITCH_() do { \
-        *Q_UINT2PTR_CAST(uint32_t, 0xE000ED04U) = (uint32_t)(1U << 28); \
-        __DSB(); \
-        __ISB(); \
+    #define QXK_CONTEXT_SWITCH_() do {                        \
+        *Q_UINT2PTR_CAST(uint32_t, 0xE000ED04U) = (1U << 28); \
+        __DSB();                                              \
+        __ISB();                                              \
     } while (false)
 
 #endif
@@ -60,12 +60,12 @@
 /* QXK interrupt entry and exit */
 #define QXK_ISR_ENTRY() ((void)0)
 
-#define QXK_ISR_EXIT()  do { \
-    QF_INT_DISABLE(); \
-    if (QXK_sched_() != (uint_fast8_t)0) { \
-        *Q_UINT2PTR_CAST(uint32_t, 0xE000ED04U) = (uint32_t)(1U << 28); \
-    } \
-    QF_INT_ENABLE(); \
+#define QXK_ISR_EXIT()  do {                                  \
+    QF_INT_DISABLE();                                         \
+    if (QXK_sched_() != 0U) {                                 \
+        *Q_UINT2PTR_CAST(uint32_t, 0xE000ED04U) = (1U << 28); \
+    }                                                         \
+    QF_INT_ENABLE();                                          \
 } while (false)
 
 /* initialization of the QXK kernel */
