+++ ports/arm-cr/qk/gnu/qk_port.h
@@ -41,69 +41,69 @@
 
 /* QK-specific Interrupt Request handler BEGIN */
 #ifdef __FPU_PRESENT
-#define QK_IRQ_BEGIN(name_) \
-    void name_(void) \
-    __attribute__ ((interrupt ("irq"))); \
+#define QK_IRQ_BEGIN(name_)                    \
+    void name_(void)                           \
+    __attribute__ ((interrupt ("irq")));       \
     __attribute__ ((naked)) void name_(void) { \
-    __asm volatile (" SUB LR, LR, #4\n" \
-    " SRSDB #31!\n" \
-    " CPS #31\n" \
-    " PUSH {R0-R3, R12}"); \
-    __asm(" FMRX R12, FPSCR\n" \
-    " STMFD SP!, {R12}\n" \
-    " FMRX R12, FPEXC\n" \
-    " STMFD SP!, {R12}\n" \
-    " FSTMDBD SP!, {D0-D7}"); \
-    __asm(" AND R3, SP, #4\n" \
-    " SUB SP, SP, R3\n" \
-    " PUSH {R3, LR}\n"); \
+    __asm volatile (" SUB LR, LR, #4\n"        \
+    " SRSDB #31!\n"                            \
+    " CPS #31\n"                               \
+    " PUSH {R0-R3, R12}");                     \
+    __asm(" FMRX R12, FPSCR\n"                 \
+    " STMFD SP!, {R12}\n"                      \
+    " FMRX R12, FPEXC\n"                       \
+    " STMFD SP!, {R12}\n"                      \
+    " FSTMDBD SP!, {D0-D7}");                  \
+    __asm(" AND R3, SP, #4\n"                  \
+    " SUB SP, SP, R3\n"                        \
+    " PUSH {R3, LR}\n");                       \
     ++QK_attr_.intNest; {
 #else
-#define QK_IRQ_BEGIN(name_) \
-    void name_(void) \
-    __attribute__ ((target ("arm"))); \
+#define QK_IRQ_BEGIN(name_)                    \
+    void name_(void)                           \
+    __attribute__ ((target ("arm")));          \
     __attribute__ ((naked)) void name_(void) { \
-    __asm volatile (" SUB LR, LR, #4\n" \
-    " SRSDB #31!\n" \
-    " CPS #31\n" \
-    " PUSH {R0-R3, R12}"); \
-    __asm(" AND R3, SP, #4\n" \
-    " SUB SP, SP, R3\n" \
-    " PUSH {R3, LR}\n"); \
+    __asm volatile (" SUB LR, LR, #4\n"        \
+    " SRSDB #31!\n"                            \
+    " CPS #31\n"                               \
+    " PUSH {R0-R3, R12}");                     \
+    __asm(" AND R3, SP, #4\n"                  \
+    " SUB SP, SP, R3\n"                        \
+    " PUSH {R3, LR}\n");                       \
     ++QK_attr_.intNest; {
 #endif
 
 /* QK-specific Interrupt Request handler END */
 #ifdef __FPU_PRESENT
-#define QK_IRQ_END() \
-    } --QK_attr_.intNest; \
-    if (QK_attr_.intNest == (uint_fast8_t)0) { \
-        if (QK_sched_() != (uint_fast8_t)0) { \
-            QK_activate_(); \
-        } \
-    } \
+#define QK_IRQ_END()                  \
+    } --QK_attr_.intNest;             \
+    if (QK_attr_.intNest == 0U) {     \
+        if (QK_sched_() != 0U) {      \
+            QK_activate_();           \
+        }                             \
+    }                                 \
     __asm volatile (" POP {R3, LR}\n" \
-    " ADD SP, SP, R3"); \
-    __asm(" FLDMIAD SP!, {D0-D7}\n" \
-    " LDMFD SP!, {R12}\n" \
-    " FMXR FPEXC, R12 \n" \
-    " LDMFD SP!, {R12} \n" \
-    " FMXR FPSCR, R12"); \
-    __asm(" POP {R0-R3, R12}\n" \
-    " RFEIA SP!"); \
+    " ADD SP, SP, R3");               \
+    __asm(" FLDMIAD SP!, {D0-D7}\n"   \
+    " LDMFD SP!, {R12}\n"             \
+    " FMXR FPEXC, R12 \n"             \
+    " LDMFD SP!, {R12} \n"            \
+    " FMXR FPSCR, R12");              \
+    __asm(" POP {R0-R3, R12}\n"       \
+    " RFEIA SP!");                    \
 }
 #else
-#define QK_IRQ_END() \
-    } --QK_attr_.intNest; \
-    if (QK_attr_.intNest == (uint_fast8_t)0) { \
-        if (QK_sched_() != (uint_fast8_t)0) { \
-            QK_activate_(); \
-        } \
-    } \
-    __asm volatile (" POP {R3, LR}\n" \
-    " ADD SP, SP, R3"); \
+#define QK_IRQ_END()                 \
+    } --QK_attr_.intNest;            \
+    if (QK_attr_.intNest == 0U) {    \
+        if (QK_sched_() != 0U) {     \
+            QK_activate_();          \
+        }                            \
+    }                                \
+    __asm volatile (" POP {R3, LR}\n"     \
+    " ADD SP, SP, R3");                   \
     __asm volatile (" POP {R0-R3, R12}\n" \
-    " RFEIA SP!"); \
+    " RFEIA SP!");                        \
 }
 #endif
